final commit
implemented saving the player name between the menu scene and main scene, did this by creating a MenuManager script with a static instance that would not be destroyed on load.

implemented saving the high score (and the playerName who made the high score) between sessions using a SaveData class and writing to a file.